Extract 'get value of type x' code from ArrayEditor and EditDialog into ValueFactory
