+++ src/UI/Editor.cs
@@ -614,7 +614,7 @@ public class Editor : Toplevel
 
         // BUG: This is an improper exception here and could have unexpected behavior if this method is ever called asynchronously.
         var factory = new OperationFactory(
-                (p, v) => EditDialog.GetNewValue(p.Design, p, v, out var newValue) ? newValue : throw new OperationCanceledException() );
+                (p, v) => ValueFactory.GetNewValue(p.Design, p, v, out var newValue) ? newValue : throw new OperationCanceledException() );
 
         var operations = factory
             .CreateOperations(selected, m, rightClicked, out string name)
